+++ src/main/java/training/endava/app/repository/jparepository/PersonDetailsRepositoryImpl.java
@@ -3,10 +3,8 @@ package training.endava.app.repository.jparepository;
 import org.springframework.stereotype.Repository;
 import training.endava.app.database.PostgresJPA;
 import training.endava.app.domain.entity.PersonDetails;
-import training.endava.app.domain.entity.PhoneNumber;
 
 import javax.persistence.EntityManager;
-import javax.persistence.EntityNotFoundException;
 import javax.persistence.TypedQuery;
 import java.util.ArrayList;
 import java.util.List;
@@ -17,6 +15,9 @@ public class PersonDetailsRepositoryImpl implements PersonDetailsRepository{
     private EntityManager em;
 
     public void add(PersonDetails personDetails) {
+        if(getById(personDetails.getId()) != null){
+            throw new RuntimeException("Person already exists in DB");
+        }
         em = PostgresJPA.getInstance();
         em.getTransaction().begin();
         em.persist(personDetails);
@@ -25,6 +26,9 @@ public class PersonDetailsRepositoryImpl implements PersonDetailsRepository{
     }
 
     public void delete(PersonDetails personDetails) {
+        if(getById(personDetails.getId()) == null){
+            throw new RuntimeException("Person doesn't exists in DB");
+        }
         em = PostgresJPA.getInstance();
         em.getTransaction().begin();
         em.remove(em.contains(personDetails) ? personDetails : em.merge(personDetails));
@@ -33,14 +37,13 @@ public class PersonDetailsRepositoryImpl implements PersonDetailsRepository{
     }
 
     public PersonDetails getById(Long id) {
+        if(id == null){
+            return null;
+        }
         PersonDetails personDetails;
         em = PostgresJPA.getInstance();
         em.getTransaction().begin();
         personDetails = em.find(PersonDetails.class, id);
-        if (personDetails == null) {
-            throw new EntityNotFoundException("Can't find Artist for ID "
-                    + id);
-        }
         em.close();
 
         return personDetails;
@@ -48,7 +51,7 @@ public class PersonDetailsRepositoryImpl implements PersonDetailsRepository{
 
     public List<PersonDetails> getAll(){
         em = PostgresJPA.getInstance();
-        TypedQuery<PersonDetails> query = em.createNamedQuery("PhoneNumber.findAll", PersonDetails.class);
+        TypedQuery<PersonDetails> query = em.createNamedQuery("PhoneDetails.findAll", PersonDetails.class);
         List<PersonDetails> personDetails = new ArrayList<>(query.getResultList());
         em.close();
 
@@ -56,6 +59,9 @@ public class PersonDetailsRepositoryImpl implements PersonDetailsRepository{
     }
 
     public void update (PersonDetails personDetails){
+        if (getById(personDetails.getId()) == null){
+            throw new RuntimeException("PersonDetails doesn't exists in DB");
+        }
         em = PostgresJPA.getInstance();
         em.getTransaction().begin();
         em.merge(personDetails);
