+++ src/main/java/training/endava/app/repository/jparepository/PhoneNumberRepositoryImp.java
@@ -1,22 +1,22 @@
 package training.endava.app.repository.jparepository;
 
 import org.springframework.stereotype.Repository;
-import training.endava.app.database.Postgres;
 import training.endava.app.database.PostgresJPA;
 import training.endava.app.domain.entity.PhoneNumber;
 
 import javax.persistence.EntityManager;
-import javax.persistence.EntityNotFoundException;
-import javax.persistence.Persistence;
 import javax.persistence.TypedQuery;
 import java.util.ArrayList;
 import java.util.List;
 
 @Repository
-public class PhoneNumberRepositoryImp implements PhoneNumberRepository{
+public class PhoneNumberRepositoryImp implements PhoneNumberRepository {
     private EntityManager em;
 
     public void add(PhoneNumber phoneNumber) {
+        if(getById(phoneNumber.getId()) != null){
+            throw new RuntimeException("PhoneNumber Already Exists");
+        }
         em = PostgresJPA.getInstance();
         em.getTransaction().begin();
         em.persist(phoneNumber);
@@ -25,6 +25,9 @@ public class PhoneNumberRepositoryImp implements PhoneNumberRepository{
     }
 
     public void delete(PhoneNumber phoneNumber) {
+        if(getById(phoneNumber.getId()) == null){
+            throw new RuntimeException("PhoneNumber doesn't exists in Db");
+        }
         em = PostgresJPA.getInstance();
         em.getTransaction().begin();
         em.remove(em.contains(phoneNumber) ? phoneNumber : em.merge(phoneNumber));
@@ -33,20 +36,19 @@ public class PhoneNumberRepositoryImp implements PhoneNumberRepository{
     }
 
     public PhoneNumber getById(Long id) {
+        if (id == null) {
+            return null;
+        }
         PhoneNumber phoneNumber;
         em = PostgresJPA.getInstance();
         em.getTransaction().begin();
         phoneNumber = em.find(PhoneNumber.class, id);
-        if (phoneNumber == null) {
-            throw new EntityNotFoundException("Can't find Artist for ID "
-                    + id);
-        }
         em.close();
 
         return phoneNumber;
     }
 
-    public List<PhoneNumber> getAll(){
+    public List<PhoneNumber> getAll() {
         em = PostgresJPA.getInstance();
         TypedQuery<PhoneNumber> query = em.createNamedQuery("PhoneNumber.findAll", PhoneNumber.class);
         List<PhoneNumber> phoneNumbers = new ArrayList<>(query.getResultList());
@@ -55,7 +57,10 @@ public class PhoneNumberRepositoryImp implements PhoneNumberRepository{
         return phoneNumbers;
     }
 
-    public void update (PhoneNumber phoneNumber){
+    public void update(PhoneNumber phoneNumber) {
+        if(getById(phoneNumber.getId()) == null){
+            throw new RuntimeException("PhoneNumber doesn't exists in DB");
+        }
         em = PostgresJPA.getInstance();
         em.getTransaction().begin();
         em.merge(phoneNumber);
